auth: backend heavy refactoring (break into submodules)
