+++ src/backend/src/auth/auth.module.ts
@@ -1,14 +1,18 @@
 import { Module } from '@nestjs/common';
-import { AuthController } from './auth.controller';
-import { JwtModule } from '@nestjs/jwt';
-import { AuthService } from './auth.service';
-import { JwtStrategy } from './jwt/strategy';
-import { OAuth2Strategy } from './oauth/strategy';
-import { HttpModule } from '@nestjs/axios';
+import { LocalAuthModule } from './local/local-auth.module';
+import { RouterModule } from '@nestjs/core';
+import { routes } from './routes';
+import { AuthUtilsModule } from './utils/auth-utils.module';
+import { JwtAuthModule } from './jwt/jwt-auth.module';
+import { OauthModule } from './oauth/oauth.module';
 
 @Module({
-  imports: [JwtModule.register({}), HttpModule],
-  controllers: [AuthController],
-  providers: [AuthService, JwtStrategy, OAuth2Strategy],
+  imports: [
+    RouterModule.register(routes),
+    LocalAuthModule,
+    JwtAuthModule,
+    AuthUtilsModule,
+    OauthModule,
+  ],
 })
 export class AuthModule {}
