+++ src/backend/src/auth/jwt/jwt-auth.module.ts
@@ -1,0 +1,11 @@
+import { Module } from '@nestjs/common';
+import { JwtModule } from '@nestjs/jwt';
+import { JwtStrategy } from './strategy';
+import { JwtAuthService } from './jwt-auth.service';
+
+@Module({
+  imports: [JwtModule.register({})],
+  providers: [JwtStrategy, JwtAuthService],
+  exports: [JwtAuthService],
+})
+export class JwtAuthModule {}
