+++ src/backend/src/auth/jwt/jwt-auth.service.spec.ts
@@ -1,0 +1,19 @@
+import { JwtService } from '@nestjs/jwt';
+import { Test, TestingModule } from '@nestjs/testing';
+import { JwtAuthService } from './jwt-auth.service';
+
+describe('JwtAuthService', () => {
+  let service: JwtAuthService;
+
+  beforeEach(async () => {
+    const module: TestingModule = await Test.createTestingModule({
+      providers: [JwtAuthService, JwtService],
+    }).compile();
+
+    service = module.get<JwtAuthService>(JwtAuthService);
+  });
+
+  it('should be defined', () => {
+    expect(service).toBeDefined();
+  });
+});
