+++ src/backend/src/auth/jwt/jwt-auth.service.ts
@@ -1,0 +1,15 @@
+import { Injectable } from '@nestjs/common';
+import { JwtService } from '@nestjs/jwt';
+import { JwtPayload, OAuthJwtPayload } from './interfaces';
+
+@Injectable()
+export class JwtAuthService {
+  constructor(private jwt: JwtService) {}
+
+  signToken(payload: JwtPayload | OAuthJwtPayload): Promise<string> {
+    return this.jwt.signAsync(payload, {
+      expiresIn: '42m',
+      secret: process.env['JWT_SECRET'],
+    });
+  }
+}
