+++ src/backend/src/auth/local/local-auth.service.spec.ts
@@ -1,0 +1,28 @@
+import { JwtService } from '@nestjs/jwt';
+import { Test, TestingModule } from '@nestjs/testing';
+import { DbService } from 'src/db/db.service';
+import { JwtAuthService } from '../jwt/jwt-auth.service';
+import { AuthUtilsService } from '../utils/auth-utils.service';
+import { LocalAuthService } from './local-auth.service';
+
+describe('LocalAuthService', () => {
+  let service: LocalAuthService;
+
+  beforeEach(async () => {
+    const module: TestingModule = await Test.createTestingModule({
+      providers: [
+        LocalAuthService,
+        DbService,
+        AuthUtilsService,
+        JwtAuthService,
+        JwtService,
+      ],
+    }).compile();
+
+    service = module.get<LocalAuthService>(LocalAuthService);
+  });
+
+  it('should be defined', () => {
+    expect(service).toBeDefined();
+  });
+});
