+++ src/backend/src/auth/oauth/oauth.service.spec.ts
@@ -1,0 +1,30 @@
+import { HttpModule } from '@nestjs/axios';
+import { JwtService } from '@nestjs/jwt';
+import { Test, TestingModule } from '@nestjs/testing';
+import { DbService } from 'src/db/db.service';
+import { JwtAuthService } from '../jwt/jwt-auth.service';
+import { AuthUtilsService } from '../utils/auth-utils.service';
+import { OauthService } from './oauth.service';
+
+describe('OauthService', () => {
+  let service: OauthService;
+
+  beforeEach(async () => {
+    const module: TestingModule = await Test.createTestingModule({
+      providers: [
+        OauthService,
+        DbService,
+        JwtAuthService,
+        JwtService,
+        AuthUtilsService,
+      ],
+      imports: [HttpModule],
+    }).compile();
+
+    service = module.get<OauthService>(OauthService);
+  });
+
+  it('should be defined', () => {
+    expect(service).toBeDefined();
+  });
+});
