+++ src/backend/src/auth/utils/auth-utils.module.ts
@@ -1,0 +1,15 @@
+import { Module } from '@nestjs/common';
+import { AuthUtilsService } from './auth-utils.service';
+import { AuthUtilsController } from './auth-utils.controller';
+import { HttpModule } from '@nestjs/axios';
+import { DbModule } from 'src/db/db.module';
+import { JwtModule } from '@nestjs/jwt';
+import { JwtAuthModule } from '../jwt/jwt-auth.module';
+
+@Module({
+  providers: [AuthUtilsService],
+  controllers: [AuthUtilsController],
+  imports: [HttpModule, DbModule, JwtModule, JwtAuthModule],
+  exports: [AuthUtilsService],
+})
+export class AuthUtilsModule {}
